// api/src/components/universes/universesModel.ts
import { ICharacter } from '../users/characters/characterModel';

export interface IUniverse {
  _id: string;
  universeName: string;
  characters: ICharacter[];
  createdAt: Date;
  updatedAt: Date;
}

export interface UniverseData {
  _id: string;
  universeName: string;
  characters: ICharacter[];
  createdAt: Date;
  updatedAt: Date;
}